Moved patch_sysv_wrapper to own file
